chore: rename to different binaries
