--- example/size.h
@@ -1,5 +1,0 @@
-#pragma once
-#include <stdint.h>
-uint32_t PQCLEAN_DILITHIUM5_CRYPTO_BYTES = 4627;
-uint32_t PQCLEAN_DILITHIUM5_CRYPTO_SECRETKEYBYTES = 4896;
-uint32_t PQCLEAN_DILITHIUM5_CRYPTO_PUBLICKEYBYTES = 2592;
